Minor changes to the existing score checking commands
